chore: path alias 설정
