+++ src/App.tsx
@@ -1,0 +1,22 @@
+import React from 'react';
+import {SafeAreaView, ScrollView, useColorScheme} from 'react-native';
+
+import {Colors} from 'react-native/Libraries/NewAppScreen';
+
+function App(): React.JSX.Element {
+  const isDarkMode = useColorScheme() === 'dark';
+
+  const backgroundStyle = {
+    backgroundColor: isDarkMode ? Colors.darker : Colors.lighter,
+  };
+
+  return (
+    <SafeAreaView style={backgroundStyle}>
+      <ScrollView
+        contentInsetAdjustmentBehavior="automatic"
+        style={backgroundStyle}></ScrollView>
+    </SafeAreaView>
+  );
+}
+
+export default App;
